General fixes, improvements

+++ Moongoose/src/Moongoose/Ui/ImGuiLayer.cpp
@@ -274,7 +274,7 @@ bool Moongoose::ImGuiLayer::onWindowResized(WindowResizeEvent& event)
 {
 	ImGuiIO& io = ImGui::GetIO();
 
-	io.DisplaySize = ImVec2(event.getWidth(), event.getHeight());
+	io.DisplaySize = ImVec2((float) event.getWidth(), (float) event.getHeight());
 	io.DisplayFramebufferScale = ImVec2(1.0f, 1.0f);
 	glViewport(0, 0, event.getWidth(), event.getHeight());
